requirements.txt
pandas
matplotlib
PyYAML
pytest
numpy
sv-ttk>=2.6.1